Add `getProps` to Parsers and fix up usages
